CREATED: CreateKeepFormModal.vue.
ADDED: App.vue - <CreateKeepFormModal />. Navbar.vue - Added Create Keep button to open CreateKeepFormModal.

REFACTORED: Fixed typo in CreateVaultFormModal.vue. HomePage.vue - moved CreateKeepForm to CreateKeepFormModal.vue.
